Added types for answers and comments Removed unused search function
